+++ cli/create.ts
@@ -50,8 +50,6 @@ export default async function createApp() {
   const link = LINK;
   const cwd = Deno.cwd();
   const dir = join(cwd, app);
-  await Deno.mkdir(join(dir, "@shared"), { recursive: true });
-  await Deno.mkdir(join(dir, "@shared", "result"));
   await Deno.mkdir(join(dir, "pages", "api"), { recursive: true });
   await Deno.mkdir(join(dir, "pages", "_default"));
   await Deno.mkdir(join(dir, ".vscode"));
@@ -74,7 +72,7 @@ export default async function createApp() {
   "fmt": {
     "files": {
       "exclude": [
-        "@shared/",
+        ".maze/",
         "public/",
         "server_prod.js"
       ]
@@ -83,7 +81,7 @@ export default async function createApp() {
   "lint": {
     "files": {
       "exclude": [
-        "@shared/",
+        ".maze/",
         "public/",
         "server_prod.js"
       ]
@@ -103,6 +101,12 @@ export default async function createApp() {
 }`,
   );
   await Deno.writeTextFile(
+    join(dir, ".gitignore"),
+    `.maze
+.netlify
+node_modules`,
+  );
+  await Deno.writeTextFile(
     join(dir, ".vscode", "settings.json"),
     `{
   "deno.enable": true,
@@ -110,6 +114,7 @@ export default async function createApp() {
   "deno.suggest.imports.hosts": {
     "https://deno.land": true,
     "https://esm.sh": true,
+    "https://jspm.dev": true,
     "https://cdn.skypack.dev": true
   },
   "deno.importMap": "./import_map.json"
@@ -118,39 +123,13 @@ export default async function createApp() {
   await Deno.writeTextFile(
     join(dir, "maze.config.ts"),
     `import type { MazeConfig } from "maze";
-import { ENV } from "./@shared/result/constant.ts";
   
 export default <MazeConfig>{ 
 
   // cache-control (production only).
-  cache_control: "public, max-age=31536000, immutable",
+  cache_control: "public, max-age=31536000, immutable"
 
-  // example set etag on production.
-  etag: ENV === 'production'
 }`,
-  );
-  await Deno.writeTextFile(
-    join(dir, "@shared", "maze.ts"),
-    `import core from "./core.ts";
-import { NHttp, ReqEvent } from "${link}/core/server.ts";
-
-export default (static_url?: string, {
-  routerCallback,
-  staticConfig
-}: {
-  routerCallback?: (router: NHttp<ReqEvent>) => any
-  staticConfig?: (rev: ReqEvent) => void
-} = {}) => core(static_url, { routerCallback, staticConfig });
-`,
-  );
-  await Deno.writeTextFile(
-    join(dir, "server.ts"),
-    `import maze from "./@shared/maze.ts";
-
-maze(import.meta.url).listen(8080, () => {
-  console.log("> Running on http://localhost:8080");
-});
-`,
   );
   await Deno.writeTextFile(
     join(dir, "pages", "_default", "app.tsx"),
@@ -244,185 +223,6 @@ export default class Home extends Component<PageProps> {
   }
 } 
 `,
-  );
-  await Deno.writeTextFile(
-    join(dir, "@shared", "root_app.tsx"),
-    `/** @jsx h */
-import { h } from "nano-jsx";
-import App from "../pages/_default/app.tsx";
-
-function RootApp({ Page, initData, route, isServer }: any) {
-  return (
-    <App
-      Page={Page}
-      props={{ ...initData, route, isServer }}
-    />
-  );
-}
-
-RootApp.initProps = (App as any).initProps;
-RootApp.event = (App as any).event || {};
-
-export default RootApp;
-`,
-  );
-  await Deno.writeTextFile(
-    join(dir, "@shared", "result", "pages.ts"),
-    `
-import $0 from "../../pages/index.tsx";
-export const pages: any = [
-  { 
-    path: '/',
-    page: $0,
-    methods: ($0 as any).methods,
-    hydrate: ($0 as any).hydrate
-  },
-];
-`,
-  );
-  await Deno.writeTextFile(
-    join(dir, "@shared", "result", "constant.ts"),
-    `export const BUILD_ID: string = '${Date.now()}';
-export const ENV: string = 'development';`,
-  );
-  await Deno.writeTextFile(
-    join(dir, "@shared", "result", "apis.ts"),
-    `
-import { Router } from "https://deno.land/x/nhttp@${NHTTP_VERSION}/mod.ts";
-import { RequestEvent } from "${link}/core/types.ts";
-const api = new Router<RequestEvent>();
-export default api;
-`,
-  );
-  await Deno.writeTextFile(
-    join(dir, "@shared", "result", "server_pages.ts"),
-    `
-import $0 from "../../pages/index.tsx";
-export const pages: any = [
-  { 
-    path: '/',
-    page: $0,
-    methods: ($0 as any).methods,
-    hydrate: ($0 as any).hydrate
-  },
-];
-`,
-  );
-  await Deno.writeTextFile(
-    join(dir, "@shared", "core.ts"),
-    `import { initApp as baseInitApp } from "${link}/core/server.ts";
-import ErrorPage from "../pages/_default/error.tsx";
-import ssr from "../pages/_default/ssr.ts";
-import config from "../maze.config.ts";
-import RootApp from "./root_app.tsx";
-import apis from "./result/apis.ts";
-import { pages } from "./result/pages.ts";
-import { BUILD_ID } from "./result/constant.ts";
-import { pages as server_pages } from "./result/server_pages.ts";
-
-export default (static_url?: string, {
-  routerCallback,
-  staticConfig,
-} = {} as any) => {
-  return baseInitApp({
-    root: RootApp,
-    error_page: ErrorPage,
-    pages: pages,
-    server_pages: server_pages,
-    apis: apis,
-    meta_url: static_url,
-    build_id: BUILD_ID,
-    ssr: ssr,
-    static_config: staticConfig,
-    etag: config.etag,
-    cache_control: config.cache_control
-  }, routerCallback);
-};
-`,
-  );
-  await Deno.writeTextFile(
-    join(dir, "@shared", "hydrate.tsx"),
-    `/** @jsx h */
-import { h } from "nano-jsx";
-import { pages } from "./result/pages.ts";
-import RootApp from "./root_app.tsx";
-import config from "../maze.config.ts";
-import { onHydrate } from "../pages/_default/client.ts";
-import ErrorPage from "../pages/_default/error.tsx";
-import ClassicRouter from "${LINK}/core/classic_router.tsx";
-
-async function lazy(url: string) {
-  const mod = (await import(url)).default;
-  return mod;
-}
-onHydrate();
-let first = true;
-let init: any = document.getElementById("__INIT_DATA__");
-if (init) init = JSON.parse(init.textContent || "{}");
-const router = new ClassicRouter(ErrorPage);
-const _pages = router.buildPages(location.pathname, (config.zones || []) as string[], pages);
-for (let i = 0; i < _pages.length; i++) {
-  const obj: any = _pages[i];
-  router.add(obj.path, async (rev) => {
-    rev.isFirst = first;
-    try {
-      let rootData = {};
-      if (!first) {
-        rootData = RootApp.initProps ? (await RootApp.initProps(rev)) : {};
-      }
-      if (RootApp.event.onStart !== void 0) {
-        await RootApp.event.onStart(rev);
-      }
-      const Page: any = typeof obj.page === "string"
-        ? (await lazy(obj.page))
-        : obj.page;
-      const initData = first
-        ? init || {}
-        : (Page.initProps ? (await Page.initProps(rev)) : {});
-      const initRender = () => {
-        rev.render(
-          <RootApp
-            Page={Page}
-            initData={{ ...initData, ...rootData }}
-            route={{
-              pathname: rev.pathname,
-              url: rev.url,
-              path: obj.path,
-              params: rev.params,
-            }}
-            isServer={false}
-          />,
-        );
-      }
-      if (first) {
-        initRender();
-      } else {
-        const myInitData = { ...initData, ...rootData };
-        rev.render(
-          <Page
-            {...myInitData}
-            route={{
-              pathname: rev.pathname,
-              url: rev.url,
-              path: obj.path,
-              params: rev.params,
-            }}
-            isServer={false}
-          />, "__MAZE_PAGE__"
-        );
-      }
-      if (RootApp.event.onEnd !== void 0) {
-        RootApp.event.onEnd(rev);
-      }
-    } catch (err) {
-      if (RootApp.event.onError !== void 0) {
-        RootApp.event.onError(err, rev);
-      }
-    }
-    first = false;
-  });
-}
-router.resolve();`,
   );
   console.log(`Success create ${app}.
