+++ core/gen.ts
@@ -175,11 +175,11 @@ export async function genPages(
     // save pages.ts
     if (stat_page.status || refresh) {
       const str_file = genRoutes(page_list, "page", env);
-      const path = join(resolve(dir, "./@shared/result/pages.ts"));
+      const path = join(resolve(dir, "./.maze/result/pages.ts"));
       await Deno.writeTextFile(path, str_file);
       if (env === "production" || env === "production_bundles") {
         const str_file = genRoutes(page_list, "page", "development");
-        const path = join(resolve(dir, "./@shared/result/server_pages.ts"));
+        const path = join(resolve(dir, "./.maze/result/server_pages.ts"));
         await Deno.writeTextFile(path, str_file);
       }
     }
@@ -187,7 +187,7 @@ export async function genPages(
     // save apis.ts
     if (stat_api.status || refresh) {
       const str_file = genRoutes(api_list, "api", env);
-      const path = join(resolve(dir, "./@shared/result/apis.ts"));
+      const path = join(resolve(dir, "./.maze/result/apis.ts"));
       await Deno.writeTextFile(path, str_file);
     }
   } catch (error) {
