+++ core/server_prod.ts
@@ -1,5 +1,5 @@
 import { NHttp } from "./deps.ts";
-import { ReqEvent, TOptionsInitApp, TRet } from "./types.ts";
+import { ReqEvent, TOptionsInitApp } from "./types.ts";
 import baseInitApp from "./init_app.tsx";
 
 const env = "production";
@@ -7,7 +7,6 @@ const app = new NHttp<ReqEvent>({ env });
 
 export const initApp = (
   opts: TOptionsInitApp,
-  routeCallback?: (app: NHttp<ReqEvent>) => TRet,
 ) => {
   const clientScript = `/__maze/${opts.build_id}/_app.js`;
   const myApp = baseInitApp(
@@ -18,7 +17,6 @@ export const initApp = (
     },
     opts.server_pages,
     app,
-    routeCallback,
   );
   return myApp;
 };
