+++ tests/sample/.maze/maze.ts
@@ -1,0 +1,3 @@
+import core from "./core.ts";
+
+export default (static_url?: string) => core(static_url);
