+++ tests/sample/.maze/result/constant.ts
