--- tests/sample/@shared/maze.ts
@@ -1,10 +1,0 @@
-import core from "./core.ts";
-import { NHttp, ReqEvent } from "../../../core/server.ts";
-
-export default (static_url?: string, {
-  routerCallback,
-  staticConfig,
-}: {
-  routerCallback?: (router: NHttp<ReqEvent>) => any;
-  staticConfig?: (rev: ReqEvent) => void;
-} = {}) => core(static_url, { routerCallback, staticConfig });
